invoice-parser/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

supabase>=2.6.0,<3.0.0
httpx==0.28.1

google-cloud-storage==2.16.0
pdfplumber==0.10.3
python-dateutil==2.8.2
requests==2.31.0

openai>=1.0.0
pypdf>=4.0.0

